Remove common test
