+++ aws-crt-kotlin/native/test/aws/sdk/kotlin/crt/io/MutableBufferTest.kt
@@ -5,10 +5,12 @@
 package aws.sdk.kotlin.crt.io
 
 import aws.sdk.kotlin.crt.CrtTest
+import kotlinx.cinterop.ExperimentalForeignApi
 import kotlin.test.Test
 import kotlin.test.assertContentEquals
 import kotlin.test.assertEquals
 
+@OptIn(ExperimentalForeignApi::class)
 class MutableBufferTest : CrtTest() {
     @Test
     fun testSimpleWrite() {
@@ -23,5 +25,17 @@ class MutableBufferTest : CrtTest() {
         assertContentEquals(data.encodeToByteArray(), buffer.bytes.copyOfRange(0, data.length))
 
         buffer.close()
+    }
+
+    @Test
+    fun testWriteToFullBuffer() {
+        val str = "Hello!"
+        val bytes = str.encodeToByteArray()
+        val buffer = MutableBuffer.of(bytes) // creates a full buffer
+
+        assertEquals(0, buffer.writeRemaining)
+
+        // since it's full, should write 0 bytes
+        assertEquals(0, buffer.write(bytes))
     }
 }
